More functionalities added
User account editing form and list of products being sold by the current user added, although there is work to be done with the API call.

Product lists got styled better,  the API call was optimized for cases when the user wants to see all products or products of a single category.

Product details screen was added and made functional, along with the history of products bought by the current user.

Next steps are the fixing of the requests for the user account editing, products bought by the user, and adding the shopping cart functionality to buy products from the app.