User data, Product Fix, and Overlay
The way the user data is obtained and stored was changed, allowing it to be readable to the account management screen for later updating.

The list of products was corrected so the product name is properly shown underneath the image, and the "add to cart" button was added on the individual product page.

Lastly, a preloader overlay has been added with every AJAX request made to the REST API, so the user knows in advance information is being retrieved from the server.